refactor(PIO-90): fix convention violations from code review
- Extract pipe commands from cli.js into src/pipe.js to bring cli.js under 1000 lines
- Add missing ListPipeSignalsRequest and PollPendingSeedRequest Form Request classes
- Inject Form Requests into PipeSignalController::index() and PipePairingController::pendingSeed()
- Remove unused \$request parameter from PipeController::destroy()
- Rename ClearExpiredPipeSessionsTest to ClearExpiredPipeJobsTest to match its actual scope
- Wire --expires-in flag through to createPipeSession (api.js + PipeController + CreatePipeSessionRequest)
