+++ app/Http/Requests/Api/CreatePipeSessionRequest.php
@@ -16,6 +16,7 @@ class CreatePipeSessionRequest extends FormRequest
         return [
             'session_id' => ['required', 'string', 'regex:/^[0-9a-f]{32,40}$/', 'unique:pipe_sessions,session_id'],
             'transfer_mode' => ['required', 'string', 'in:relay,p2p'],
+            'expires_in' => ['nullable', 'integer', 'min:60', 'max:3600'],
         ];
     }
 }
