+++ app/Http/Requests/Api/ListPipeSignalsRequest.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Http\Requests\Api;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class ListPipeSignalsRequest extends FormRequest
+{
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    public function rules(): array
+    {
+        return [
+            'role' => ['required', 'string', 'in:sender,receiver'],
+            'after' => ['nullable', 'integer', 'min:0'],
+        ];
+    }
+}
