+++ app/Http/Requests/Api/PollPendingSeedRequest.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Http\Requests\Api;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class PollPendingSeedRequest extends FormRequest
+{
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    public function rules(): array
+    {
+        return [
+            'device_id' => ['required', 'string'],
+        ];
+    }
+}
